FIxed some property types.

+++ Headers/UXImageView.h
@@ -35,7 +35,7 @@
 @property(retain, nonatomic) NSString *accessibilityLabel;
 @property(readonly, copy) NSString *debugDescription;
 @property(readonly, copy) NSString *description;
-@property(readonly) unsigned long long hash;
+@property(readonly) NSUInteger hash;
 @property(readonly) Class superclass;
 
 @end
